Documentation for commit d8eb43a80ae2f8fed4946d70fb2c61fabeefc67f

+++ noises/sodium/ffi/sidebar-items.js
@@ -1,1 +1,1 @@
-initSidebarItems({"enum":[["c_void","Type used to construct void pointers for use with C."]],"type":[["c_int",""],["size_t",""]]});
+initSidebarItems({"enum":[["c_void","Type used to construct void pointers for use with C."]],"type":[["c_int",""],["c_ulonglong",""]]});
